LIBFCREPO-1617. Allow substitutions in the config files.
Use `envsubst` from `plastron.utils` to allow for substitutions in the config files loaded from the `*_FILE` environment variables. The mapping of values that can be substituted is the applications current `config`, so it has access to any environment variables prefixed with `SOLRIZER_` in their non-prefixed form.

Renamed `proxy_base_url` to `proxy_prefix` to better align with naming in other systems.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1617